Changed color scheme to match MoodMusic
Note that since the SpotifyLoginButton is not an open class, it is not able to be extended. As such, I had to essentially copy and paste the code in order to modify the colors. Simply just changing the button color in the LogInViewController won't work because the SpotifyLoginButton has a method inside of it that changes the button color as a result of button touching up and down.